fix(mail): make test code compile — 555/573 mail tests now pass
The pre-flood baseline's #[cfg(test)] modules had drifted and never compiled
(masked by the bloat's syntax error). Verified on a toolchain host:
`cargo test -p expresso-mail --bins` now builds and runs — 555 pass, 18 fail.

Test-compile fixes (no production behavior change):
- QuotaDto/PresetBody/SnoozeBody: add the serde derive the round-trip tests need.
- session.rs tests: import `list_matches` (test added without updating `use super`).
- ingest.rs tests: normalize_message_id now takes Option<String> — wrap args.
- drafts.rs tests: bind build_raw() result before borrowing (E0716 temporaries).
- lmtp.rs test: MAX_RCPTS is usize, not u64.
- submission.rs test: decode_plain returns Option — is_none(), not is_err().

The 18 remaining failures are PRE-EXISTING behavioral drift, not regressions and
not in the features added this session (signatures/threads/undo-send all pass):
build_raw rejects recipient-less drafts (lettre envelope requires a recipient —
a real draft-semantics bug), plus assorted stale-assertion tests (quota 507
status, list_matches `%`, mime navigate, lmtp bracket stripping, smtp metric
mapping). Tracked for follow-up.

+++ services/expresso-mail/src/lmtp.rs
@@ -335,7 +335,7 @@ mod tests {
 
     #[test]
     fn max_msg_bytes_exceeds_max_rcpts() {
-        assert!(super::MAX_MSG_BYTES > super::MAX_RCPTS as u64);
+        assert!(super::MAX_MSG_BYTES > super::MAX_RCPTS as usize);
     }
 
     #[test]
